Criando Camadas de Pagamento e finalizando o Teste

+++ src/main/java/com/srm/empresax/emprestimos/domain/service/PessoaService.java
@@ -5,6 +5,7 @@ import com.srm.empresax.emprestimos.domain.repository.PessoaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -23,20 +24,20 @@ public class PessoaService {
 
         if (tamanho == 11) {
             pessoa.setTipoIdentificador("PF");
-            pessoa.setValorMinimoParcela(300.00);
-            pessoa.setValorMaximoEmprestimo(10000.00);
+            pessoa.setValorMinimoParcela(BigDecimal.valueOf(300.00));
+            pessoa.setValorMaximoEmprestimo(BigDecimal.valueOf(10000.00));
         } else if (tamanho == 14) {
             pessoa.setTipoIdentificador("PJ");
-            pessoa.setValorMinimoParcela(1000.00);
-            pessoa.setValorMaximoEmprestimo(100000.00);
+            pessoa.setValorMinimoParcela(BigDecimal.valueOf(1000.00));
+            pessoa.setValorMaximoEmprestimo(BigDecimal.valueOf(100000.00));
         } else if (tamanho == 8) {
             pessoa.setTipoIdentificador("EU");
-            pessoa.setValorMinimoParcela(100.00);
-            pessoa.setValorMaximoEmprestimo(10000.00);
+            pessoa.setValorMinimoParcela(BigDecimal.valueOf(100.00));
+            pessoa.setValorMaximoEmprestimo(BigDecimal.valueOf(10000.00));
         } else if (tamanho == 10) {
             pessoa.setTipoIdentificador("AP");
-            pessoa.setValorMinimoParcela(400.00);
-            pessoa.setValorMaximoEmprestimo(25000.00);
+            pessoa.setValorMinimoParcela(BigDecimal.valueOf(400.00));
+            pessoa.setValorMaximoEmprestimo(BigDecimal.valueOf(25000.00));
         } else {
             throw new IllegalArgumentException("Identificador inválido: tamanho não reconhecido.");
         }
